+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * PayPal
   * Fix an issue where `PayPalRequest` was sending `phone_number` instead of `payer_phone`
+* GooglePay
+    * Add `softwareInfo` details to the `merchantInfo` field in the `loadPaymentData` request.
 
 ## 5.15.0 (2025-08-13)